Move and split wimp: save wimp part and separate popup
* use backward compatibility imports in module
