+++ tox.ini
@@ -24,11 +24,10 @@ deps =
 commands =
   isort \
     urwid/event_loop \
+    urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py \
-    urwid/widget.py \
-    urwid/wimp.py
+    urwid/command_map.py
 
 [testenv:black]
 skip_install = true
@@ -38,11 +37,10 @@ deps =
 commands =
   black \
     urwid/event_loop \
+    urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py \
-    urwid/widget.py \
-    urwid/wimp.py
+    urwid/command_map.py
 
 [testenv:ruff]
 skip_install = true
@@ -50,8 +48,7 @@ depends = black,isort
 deps = ruff
 commands = ruff \
     urwid/event_loop \
+    urwid/widget \
     urwid/font.py \
     urwid/escape.py \
-    urwid/command_map.py \
-    urwid/widget.py \
-    urwid/wimp.py
+    urwid/command_map.py
